sample-apps/python-sample-app/requirements.txt
backoff==1.11.1
boto3==1.24.40
botocore==1.27.40
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Deprecated==1.2.13
docopt==0.6.2
Flask==2.1.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
opentelemetry-api==1.12.0rc2
opentelemetry-distro==0.32b0
opentelemetry-exporter-otlp==1.12.0rc2
opentelemetry-exporter-otlp-proto-grpc==1.12.0rc2
opentelemetry-exporter-otlp-proto-http==1.12.0rc2
opentelemetry-instrumentation==0.32b0
opentelemetry-instrumentation-aws-lambda==0.32b0
opentelemetry-instrumentation-dbapi==0.32b0
opentelemetry-instrumentation-flask==0.32b0
opentelemetry-instrumentation-jinja2==0.32b0
opentelemetry-instrumentation-logging==0.32b0
opentelemetry-instrumentation-requests==0.32b0
opentelemetry-instrumentation-sqlite3==0.32b0
opentelemetry-instrumentation-urllib==0.32b0
opentelemetry-instrumentation-wsgi==0.32b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.12.0rc2
opentelemetry-sdk==1.12.0rc2
opentelemetry-semantic-conventions==0.32b0
opentelemetry-util-http==0.32b0
pipreqs==0.4.11
protobuf==3.20.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
typing-extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
wrapt==1.14.1
yarg==0.1.9
zipp==3.8.1
